Refactoring a bit and fixing issues with selenium not staying running

+++ provisioning/sitespeed.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 set -e
 
-if [ ! -d /etc/sitespeed_provisioned_at ]
+if [ ! -f /etc/sitespeed_provisioned_at ]
   then
   # Add Google public key to apt, needed to get Chrome
   wget -q -O - https://dl-ssl.google.com/linux/linux_signing_key.pub | sudo apt-key add -
@@ -40,17 +40,19 @@ if [ ! -d /etc/sitespeed_provisioned_at ]
   # Set locale
   echo "export LC_ALL='en_US.utf8'" >> ~/.bashrc
 
-  # echo some info on login
-  echo "sitespeed.io --version" >> ~/.bashrc
-  echo "chromedriver --version" >> ~/.bashrc
-  echo "google-chrome-stable --version" >> ~/.bashrc
-  # Starting Firefox will get us this message
-  # GLib-CRITICAL **: g_slice_set_config: assertion 'sys_page_size == 0' failed
-  # https://bugzilla.mozilla.org/show_bug.cgi?id=833117
-  # echo "firefox -version"
+  # Turn off chrome auto update
+  if [ -f /etc/apt/sources.list.d/google-chrome.list ]
+    then
+    mv /etc/apt/sources.list.d/google-chrome.list /etc/apt/sources.list.d/google-chrome.list.save
+  fi
 
-  # and then turn off chrome auto update
-  mv /etc/apt/sources.list.d/google-chrome.list /etc/apt/sources.list.d/google-chrome.list.save
+  # Install Selenium
+  npm install selenium-standalone@latest -g
+  selenium-standalone install
+  cp /home/vagrant/provisioning-files/selenium-start.sh /bin/
+  chown +x /bin/selenium-start.sh
+  /bin/selenium-start.sh
+  cp /home/vagrant/provisioning-files/selenium.conf /etc/init/
 
   echo 'Installation finished'
